moved function to its proper place
