+++ src/main/kotlin/io/github/mojira/arisa/infrastructure/jira/Mappers.kt
@@ -325,29 +325,3 @@ private fun JiraIssue.getOtherUpdateContext(
                 transition()
         ).also { cache.add(key, it) }
     }
-
-private fun JiraIssue.getOtherIssue(
-    jiraClient: JiraClient,
-    messages: HelperMessages,
-    config: Config,
-    cache: IssueUpdateContextCache,
-    oldPostedCommentCache: Cache<MutableSet<String>>,
-    newPostedCommentCache: Cache<MutableSet<String>>,
-    key: String
-) : Either<Throwable, Issue> {
-    val newJiraIssue = getIssue(jiraClient, key)
-    return newJiraIssue.fold(
-        { it.left() },
-        {
-            it.toDomain(
-                jiraClient,
-                jiraClient.getProject(it.project.key),
-                messages,
-                config,
-                cache,
-                oldPostedCommentCache,
-                newPostedCommentCache
-            ).right()
-        }
-    )
-}
